refactor: move ScenarioEventService from API route directory to server layer (#1537)
Move service-layer files out of the Next.js route directory into
src/server/scenarios/ to match the standard layering convention used by
every other major API (traces, prompts, dataset). Fixes the dependency
inversion where server-layer code imported from the route directory.
