+++ langwatch/src/components/simulations/SetCard.tsx
@@ -1,6 +1,6 @@
 import { Box, Card, HStack, Text, VStack } from "@chakra-ui/react";
 import { Settings } from "lucide-react";
-import type { ScenarioSetData } from "~/app/api/scenario-events/[[...route]]/types";
+import type { ScenarioSetData } from "~/server/scenarios/scenario-event.types";
 import {
   isOnPlatformSet,
   ON_PLATFORM_DISPLAY_NAME,
