+++ langwatch/src/components/simulations/SimulationCard.tsx
@@ -1,5 +1,5 @@
 import { Badge, Box, Card, HStack, Text, VStack } from "@chakra-ui/react";
-import { ScenarioRunStatus } from "~/app/api/scenario-events/[[...route]]/enums";
+import { ScenarioRunStatus } from "~/server/scenarios/scenario-event.enums";
 import { SimulationStatusOverlay } from "./SimulationStatusOverlay";
 
 export interface SimulationCardMessage {
