+++ langwatch/src/components/simulations/set-run-history-sidebar/RunHistoryItem.tsx
@@ -1,6 +1,6 @@
 import { Box, HStack, Text } from "@chakra-ui/react";
 import React from "react";
-import { ScenarioRunStatus } from "~/app/api/scenario-events/[[...route]]/enums";
+import { ScenarioRunStatus } from "~/server/scenarios/scenario-event.enums";
 import { ScenarioRunStatusIcon } from "~/components/simulations/ScenarioRunStatusIcon";
 import { useSimulationRouter } from "~/hooks/simulations/useSimulationRouter";
 import type { RunItem } from "./types";
