+++ langwatch/src/components/simulations/set-run-history-sidebar/types.ts
@@ -1,4 +1,4 @@
-import type { ScenarioRunStatus } from "~/app/api/scenario-events/[[...route]]/enums";
+import type { ScenarioRunStatus } from "~/server/scenarios/scenario-event.enums";
 
 // Types for props
 export type RunItem = {
