+++ langwatch/src/components/simulations/simulation-console/CriteriaDetails.tsx
@@ -1,5 +1,5 @@
 import { Box, Text, VStack } from "@chakra-ui/react";
-import type { ScenarioResults } from "~/app/api/scenario-events/[[...route]]/schemas";
+import type { ScenarioResults } from "~/server/scenarios/schemas";
 import { CONSOLE_COLORS, REASONING_VERDICT_COLOR_MAP } from "./constants";
 
 interface CriteriaDetailsProps {
