+++ langwatch/src/components/simulations/utils/__tests__/convert-scenario-messages.unit.test.ts
@@ -1,6 +1,6 @@
 import { Role } from "@copilotkit/runtime-client-gql";
 import { describe, expect, it } from "vitest";
-import type { ScenarioMessageSnapshotEvent } from "~/app/api/scenario-events/[[...route]]/types";
+import type { ScenarioMessageSnapshotEvent } from "~/server/scenarios/scenario-event.types";
 import { convertScenarioMessagesToCopilotKit } from "../convert-scenario-messages";
 
 describe("convertScenarioMessagesToCopilotKit", () => {
