+++ langwatch/src/components/simulations/utils/convert-scenario-messages.ts
@@ -7,7 +7,7 @@ import {
   Role,
   TextMessage,
 } from "@copilotkit/runtime-client-gql";
-import type { ScenarioMessageSnapshotEvent } from "~/app/api/scenario-events/[[...route]]/types";
+import type { ScenarioMessageSnapshotEvent } from "~/server/scenarios/scenario-event.types";
 import { safeJsonParseOrStringFallback } from "./safe-json-parse-or-string-fallback";
 
 /**
