+++ langwatch/src/utils/pollForScenarioRun.ts
@@ -1,4 +1,4 @@
-import { ScenarioRunStatus } from "~/app/api/scenario-events/[[...route]]/enums";
+import { ScenarioRunStatus } from "~/server/scenarios/scenario-event.enums";
 import { createLogger } from "./logger";
 
 const logger = createLogger("pollForScenarioRun");
